chore: mv ConvertToDfError to dferrors
