+++ client/daemon/peer/peertask_conductor.go
@@ -47,7 +47,6 @@ import (
 	"d7y.io/dragonfly/v2/pkg/container/ring"
 	"d7y.io/dragonfly/v2/pkg/digest"
 	"d7y.io/dragonfly/v2/pkg/idgen"
-	"d7y.io/dragonfly/v2/pkg/rpc"
 	"d7y.io/dragonfly/v2/pkg/rpc/common"
 	schedulerclient "d7y.io/dragonfly/v2/pkg/rpc/scheduler/client"
 	"d7y.io/dragonfly/v2/pkg/source"
@@ -794,7 +793,7 @@ func (pt *peerTaskConductor) confirmReceivePeerPacketError(err error) (cont bool
 		failedReason string
 	)
 	// extract DfError for grpc status
-	err = rpc.ConvertToDfError(err)
+	err = dferrors.ConvertToDfError(err)
 	de, ok := err.(*dferrors.DfError)
 	if ok {
 		switch de.Code {
