+++ pkg/rpc/interceptor.go
@@ -126,7 +126,7 @@ func convertServerError(err error) error {
 // ConvertErrorUnaryClientInterceptor returns a new unary client interceptor that convert error when trigger custom error.
 func ConvertErrorUnaryClientInterceptor(ctx context.Context, method string, req, reply any, cc *grpc.ClientConn, invoker grpc.UnaryInvoker, opts ...grpc.CallOption) error {
 	if err := invoker(ctx, method, req, reply, cc, opts...); err != nil {
-		return ConvertToDfError(err)
+		return dferrors.ConvertToDfError(err)
 	}
 
 	return nil
@@ -136,23 +136,8 @@ func ConvertErrorUnaryClientInterceptor(ctx context.Context, method string, req,
 func ConvertErrorStreamClientInterceptor(ctx context.Context, desc *grpc.StreamDesc, cc *grpc.ClientConn, method string, streamer grpc.Streamer, opts ...grpc.CallOption) (grpc.ClientStream, error) {
 	s, err := streamer(ctx, desc, cc, method, opts...)
 	if err != nil {
-		return nil, ConvertToDfError(err)
+		return nil, dferrors.ConvertToDfError(err)
 	}
 
 	return s, nil
-}
-
-// ConvertToDfError converts custom error of client.
-func ConvertToDfError(err error) error {
-	for _, d := range status.Convert(err).Details() {
-		switch internal := d.(type) {
-		case *commonv1.GrpcDfError:
-			return &dferrors.DfError{
-				Code:    internal.Code,
-				Message: internal.Message,
-			}
-		}
-	}
-
-	return err
 }
